requirements.txt
torch
sentence-transformers
ollama
fastapi
uvicorn
pydantic
PyPDF2
scikit-learn
faiss-cpu